Remove unused "Sparky Solo"
