+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RoverRuckus/Deployers/ComputerTeleOp.java
@@ -4,7 +4,6 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.teamcode.RoverRuckus.BaseTeleOp;
 import org.firstinspires.ftc.teamcode.RoverRuckus.Mappings.ComputerMapping;
-import org.firstinspires.ftc.teamcode.RoverRuckus.Mappings.SoloMapping;
 
 @TeleOp(name="Sparky Terminal", group="zzz")
 public class ComputerTeleOp extends BaseTeleOp {
