+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/RoverRuckus/Deployers/TandemTeleOp.java
@@ -3,7 +3,6 @@ package org.firstinspires.ftc.teamcode.RoverRuckus.Deployers;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.teamcode.RoverRuckus.BaseTeleOp;
-import org.firstinspires.ftc.teamcode.RoverRuckus.Mappings.SoloMapping;
 import org.firstinspires.ftc.teamcode.RoverRuckus.Mappings.TandemMapping;
 @TeleOp(name="Sparky Tandem")
 public class TandemTeleOp extends BaseTeleOp {
